Refactor Slack integration to support two clients simultaneously
