+++ echo-notifications/src/main/groovy/com/netflix/spinnaker/echo/config/SlackConfig.groovy
@@ -17,72 +17,40 @@
 
 package com.netflix.spinnaker.echo.config
 
+import com.netflix.spinnaker.echo.slack.SlackAppService
 import com.netflix.spinnaker.echo.slack.SlackClient
 import com.netflix.spinnaker.echo.slack.SlackService
-import com.netflix.spinnaker.kork.web.exceptions.InvalidRequestException
 import com.netflix.spinnaker.retrofit.Slf4jRetrofitLogger
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import org.apache.commons.lang3.StringUtils
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
-import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.stereotype.Component
 import retrofit.Endpoint
 import retrofit.RestAdapter
 import retrofit.client.Client
-import retrofit.client.Response
 import retrofit.converter.JacksonConverter
-import retrofit.http.Body
-import retrofit.http.POST
-import retrofit.http.Path
 
 import static retrofit.Endpoints.newFixedEndpoint
 
 @Configuration
 @ConditionalOnProperty('slack.enabled')
+@EnableConfigurationProperties([SlackLegacyProperties, SlackAppProperties])
 @Slf4j
 @CompileStatic
 class SlackConfig {
-  final static String SLACK_INCOMING_WEBHOOK = 'https://hooks.slack.com/services'
-  final static String SLACK_CHAT_API = 'https://slack.com'
 
-  @ConfigurationProperties(prefix = "slack")
-  @Component
-  class SlackProperties {
-    private String _baseUrl
-    String token
-    String verificationToken
-    String signingSecret
-    boolean forceUseIncomingWebhook = false
-    boolean sendCompactMessages = false
-
-    boolean getUseIncomingWebhook() {
-      return forceUseIncomingWebhook || isIncomingWebhookToken(token)
-    }
-
-    void setUseIncomingWebHook() { }
-
-    boolean isIncomingWebhookToken(String token) {
-      return (StringUtils.isNotBlank(token) && token.count("/") >= 2)
-    }
-
-    void setBaseUrl(String baseUrl) {
-      this._baseUrl = baseUrl
-    }
-
-    String getBaseUrl() {
-      if (StringUtils.isNotBlank(_baseUrl)) {
-        return _baseUrl
-      } else {
-        return useIncomingWebhook ? SLACK_INCOMING_WEBHOOK : SLACK_CHAT_API;
-      }
-    }
-  }
-
+  /**
+   * This bean is used for integrations with old-style "custom integration" Slack bots, which do not support features
+   * like interactive notifications. See {@link #slackAppService} for more details.
+   *
+   * Slack documentation: https://api.slack.com/custom-integrations
+   */
   @Bean
-  SlackService slackService(SlackProperties config,
+  @Qualifier("slackLegacyService")
+  SlackService slackService(@Qualifier("slackLegacyConfig") SlackLegacyProperties config,
                             Client retrofitClient,
                             RestAdapter.LogLevel retrofitLogLevel) {
 
@@ -98,25 +66,34 @@ class SlackConfig {
         .build()
         .create(SlackClient.class)
 
-    log.info("Slack service loaded")
+    log.info("Slack legacy service loaded")
     new SlackService(slackClient, config)
   }
 
+  /**
+   * This bean is used for new-style Slack apps, which support interactive messages and other features.
+   * You can use the same token for both the legacy service above and this one (whose configuration resides
+   * in the {@code slack.app} sub-key in the config), if your integration already uses a new app, or a different
+   * token for the legacy service and this one, which might useful for migrations.
+   *
+   * Calls to {@code POST /notifications} rely on the legacy service.
+   * Calls to {@code POST /interactive-notifications} rely on the new app service.
+   */
   @Bean
-  SlackHookService slackHookService(Client retrofitClient,
-                                    RestAdapter.LogLevel retrofitLogLevel) {
-    new RestAdapter.Builder()
-      .setEndpoint(newFixedEndpoint(SLACK_INCOMING_WEBHOOK.replace("/services", "")))
+  @Qualifier("slackAppService")
+  SlackAppService slackAppService(@Qualifier("slackAppConfig") SlackAppProperties config,
+                                  Client retrofitClient,
+                                  RestAdapter.LogLevel retrofitLogLevel) {
+    def slackClient = new RestAdapter.Builder()
+      .setEndpoint(newFixedEndpoint(config.baseUrl))
       .setClient(retrofitClient)
       .setLogLevel(retrofitLogLevel)
-      .setLog(new Slf4jRetrofitLogger(SlackHookService.class))
+      .setLog(new Slf4jRetrofitLogger(SlackClient.class))
       .setConverter(new JacksonConverter())
       .build()
-      .create(SlackHookService.class)
-  }
+      .create(SlackClient.class)
 
-  interface SlackHookService {
-    @POST('/{path}')
-    Response respondToMessage(@Path(value = "path", encode = false) path, @Body Map content)
+    log.info("Slack app service loaded")
+    new SlackAppService(slackClient, config)
   }
 }
