+++ echo-notifications/src/main/groovy/com/netflix/spinnaker/echo/controller/NotificationController.groovy
@@ -21,16 +21,9 @@ import com.netflix.spinnaker.echo.notification.InteractiveNotificationCallbackHa
 import com.netflix.spinnaker.echo.notification.NotificationService
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.http.HttpHeaders
-import org.springframework.http.MediaType
-import org.springframework.http.RequestEntity
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
 @RequestMapping("/notifications")
@@ -40,34 +33,8 @@ class NotificationController {
   @Autowired(required=false)
   Collection<NotificationService> notificationServices
 
-  @Autowired
-  InteractiveNotificationCallbackHandler callbackHandler
-
   @RequestMapping(method = RequestMethod.POST)
   EchoResponse create(@RequestBody Notification notification) {
     notificationServices?.find { it.supportsType(notification.notificationType) }?.handle(notification)
   }
-
-  /**
-   * Provides a generic callback API for notification services to call, primarily in response to interactive user
-   * action (e.g. clicking a button in a message). This method makes as few assumptions as possible about the request,
-   * delegating the raw request headers, parameters and body to the corresponding [InteractiveNotificationService]
-   * to process, and similarly allows the notification service to return an arbitrary response body to the caller.
-   *
-   * @param source The unique ID of the calling notification service (e.g. "slack")
-   * @param headers The request headers
-   * @param rawBody The raw body of the request
-   * @param parameters The request parameters, parsed as a Map
-   * @return
-   */
-  @RequestMapping(
-    value = '/callbacks/{source}',
-    method = RequestMethod.POST,
-    consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-    produces = MediaType.APPLICATION_JSON_VALUE
-  )
-  ResponseEntity<String> processCallback(@PathVariable String source, RequestEntity<String> request) {
-    return callbackHandler.processCallback(source, request)
-  }
-
 }
