+++ echo-notifications/src/main/groovy/com/netflix/spinnaker/echo/notification/NotificationService.groovy
@@ -29,7 +29,8 @@ interface NotificationService {
 
 /**
  * Extension of the {@link NotificationService} interface that defines methods for services supporting interactive
- * notifications by means of callbacks from the corresponding external service into echo.
+ * notifications by means of callbacks from the corresponding external service into echo
+ * (via {@code POST /interactive-notifications/callbacks}).
  *
  * Processing of those callbacks happens in three steps:
  *
