+++ echo-notifications/src/test/groovy/com/netflix/spinnaker/echo/config/SlackConfigSpec.groovy
@@ -1,25 +1,66 @@
 package com.netflix.spinnaker.echo.config
 
+import com.netflix.spinnaker.echo.slack.SlackAppService
+import com.netflix.spinnaker.echo.slack.SlackService
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Qualifier
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.boot.test.mock.mockito.MockBean
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import retrofit.RestAdapter.LogLevel
+import retrofit.client.Client
 import spock.lang.Specification
 import spock.lang.Subject
 
+@SpringBootTest(
+  classes = [ SlackConfig, MockRetrofitConfig ],
+  properties = [
+    "slack.enabled = true",
+    // Used for the old bot
+    "slack.token = xoxb-token-for-old-bot",
+    // Used for the new bot
+    "slack.app.token = xoxb-token-for-new-bot",
+    "slack.app.verification_token = verification-token",
+    "slack.app.signing_secret = signing-secret"
+  ],
+  webEnvironment = SpringBootTest.WebEnvironment.NONE
+)
 class SlackConfigSpec extends Specification {
   @Subject
-  SlackConfig.SlackProperties configProperties
+  @Autowired
+  @Qualifier("slackLegacyConfig")
+  SlackLegacyProperties legacyConfig
 
-  def setup() {
-    configProperties = new SlackConfig.SlackProperties()
+  @Subject
+  @Autowired
+  @Qualifier("slackAppConfig")
+  SlackAppProperties appConfig
+
+  @Subject
+  @Autowired
+  @Qualifier("slackLegacyService")
+  SlackService slackLegacyService
+
+  @Subject
+  @Autowired
+  @Qualifier("slackAppService")
+  SlackAppService slackAppService
+
+  def 'legacy config loads as expected'() {
+    expect:
+    legacyConfig.token == "xoxb-token-for-old-bot"
   }
 
   def 'test slack incoming web hook is inferred correctly'() {
     given:
 
     when:
-    configProperties.token = token
+    legacyConfig.token = token
 
     then:
-    configProperties.useIncomingWebhook == expectedUseIncomingWebHook
-    configProperties.baseUrl == expectedEndpoint
+    legacyConfig.useIncomingWebhook == expectedUseIncomingWebHook
+    legacyConfig.baseUrl == expectedEndpoint
 
     where:
     token                                          | expectedEndpoint                   | expectedUseIncomingWebHook
@@ -34,12 +75,12 @@ class SlackConfigSpec extends Specification {
     given:
 
     when:
-    configProperties.token = token
-    configProperties.baseUrl = baseUrl
+    legacyConfig.token = token
+    legacyConfig.baseUrl = baseUrl
 
     then:
-    configProperties.useIncomingWebhook == expectedUseIncomingWebHook
-    configProperties.baseUrl == expectedEndpoint
+    legacyConfig.useIncomingWebhook == expectedUseIncomingWebHook
+    legacyConfig.baseUrl == expectedEndpoint
 
     where:
     token                                          | baseUrl               | expectedEndpoint                   | expectedUseIncomingWebHook
@@ -55,13 +96,13 @@ class SlackConfigSpec extends Specification {
     given:
 
     when:
-    configProperties.token = token
-    configProperties.forceUseIncomingWebhook = true
-    configProperties.baseUrl = 'https://example.com'
+    legacyConfig.token = token
+    legacyConfig.forceUseIncomingWebhook = true
+    legacyConfig.baseUrl = 'https://example.com'
 
     then:
-    configProperties.useIncomingWebhook == expectedUseIncomingWebHook
-    configProperties.baseUrl == expectedEndpoint
+    legacyConfig.useIncomingWebhook == expectedUseIncomingWebHook
+    legacyConfig.baseUrl == expectedEndpoint
 
     where:
     token                                          | expectedEndpoint      | expectedUseIncomingWebHook
@@ -70,5 +111,28 @@ class SlackConfigSpec extends Specification {
     "OLD/FASHION"                                  | "https://example.com" | true
     ""                                             | "https://example.com" | true
     null                                           | "https://example.com" | true
+  }
+
+  def 'new app config loads as expected'() {
+    expect:
+    appConfig.token == "xoxb-token-for-new-bot"
+    appConfig.verificationToken == "verification-token"
+    appConfig.signingSecret == "signing-secret"
+  }
+
+  def 'legacy and new app services use different configs and clients'() {
+    expect:
+    slackAppService.config != slackLegacyService.config
+    slackAppService.slackClient != slackLegacyService.slackClient
+  }
+
+}
+
+@Configuration
+class MockRetrofitConfig extends Specification {
+  @MockBean Client client
+
+  @Bean LogLevel getRetrofitLogLevel() {
+    return LogLevel.BASIC
   }
 }
