+++ echo-notifications/src/test/groovy/com/netflix/spinnaker/echo/notification/NotificationControllerSpec.groovy
@@ -16,29 +16,15 @@
 
 package com.netflix.spinnaker.echo.notification
 
-import com.jakewharton.retrofit.Ok3Client
 import com.netflix.spinnaker.echo.api.Notification
-import com.netflix.spinnaker.echo.api.Notification.InteractiveActionCallback
 import com.netflix.spinnaker.echo.controller.NotificationController
 import com.netflix.spinnaker.echo.notification.InteractiveNotificationCallbackHandler.SpinnakerService
-import com.netflix.spinnaker.kork.web.exceptions.InvalidRequestException
-import org.springframework.core.env.Environment
-import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpMethod
-import org.springframework.http.HttpStatus
-import org.springframework.http.RequestEntity
-import org.springframework.http.ResponseEntity
-import retrofit.client.Response
-import retrofit.mime.TypedByteArray
 import spock.lang.Specification
 import spock.lang.Subject
 
-import static java.util.Collections.emptyList
-
 class NotificationControllerSpec extends Specification {
   SpinnakerService spinnakerService
-  InteractiveNotificationService notificationService
-  InteractiveNotificationCallbackHandler callbackHandler
+  NotificationService notificationService
 
   @Subject
   NotificationController notificationController
@@ -46,15 +32,8 @@ class NotificationControllerSpec extends Specification {
   void setup() {
     notificationService = Mock()
     spinnakerService = Mock()
-    callbackHandler = new InteractiveNotificationCallbackHandler(
-      new Ok3Client(),
-      [ notificationService ],
-      Mock(Environment),
-      [ "test": spinnakerService ]
-    )
     notificationController = new NotificationController(
-      notificationServices: [ notificationService ],
-      callbackHandler: callbackHandler
+      notificationServices: [ notificationService ]
     )
   }
 
@@ -70,71 +49,5 @@ class NotificationControllerSpec extends Specification {
 
     then:
     1 * notificationService.handle(notification)
-  }
-
-  void 'an incoming callback from the notification service delegates to the appropriate service class'() {
-    given:
-    RequestEntity<String> request = new RequestEntity<>(
-      "blah", new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    InteractiveActionCallback callbackObject = new InteractiveActionCallback()
-    callbackObject.serviceId = "test"
-    callbackObject.user = "john.doe"
-
-    notificationService.supportsType(Notification.Type.SLACK) >> true
-    spinnakerService.notificationCallback(*_) >> { mockResponse() }
-
-    when:
-    notificationController.processCallback("slack", request)
-
-    then:
-    1 * notificationService.parseInteractionCallback(request) >> callbackObject
-    1 * notificationService.respondToCallback(request) >> { Optional.empty() }
-  }
-
-  void 'handling of the incoming callback throws an exception if Spinnaker service config not found'() {
-    given:
-    RequestEntity<String> request = new RequestEntity<>(
-      "blah", new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    InteractiveActionCallback callbackObject = new InteractiveActionCallback()
-    callbackObject.serviceId = "unknown"
-    callbackObject.user = "john.doe"
-
-    notificationService.supportsType(Notification.Type.SLACK) >> true
-    notificationService.parseInteractionCallback(request) >> callbackObject
-
-    when:
-    notificationController.processCallback("slack", request)
-
-    then:
-    thrown(InvalidRequestException)
-  }
-
-  void 'handling of the incoming callback returns the response generated by the corresponding service class'() {
-    given:
-    RequestEntity<String> request = new RequestEntity<>(
-      "blah", new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    ResponseEntity<String> expectedResponse = new ResponseEntity(HttpStatus.OK)
-
-    InteractiveActionCallback callbackObject = new InteractiveActionCallback()
-    callbackObject.serviceId = "test"
-    callbackObject.user = "john.doe"
-
-    notificationService.supportsType(Notification.Type.SLACK) >> true
-    spinnakerService.notificationCallback(*_) >> { mockResponse() }
-
-    when:
-    ResponseEntity<String> response = notificationController.processCallback("slack",request)
-
-    then:
-    1 * notificationService.parseInteractionCallback(request) >> callbackObject
-    1 * notificationService.respondToCallback(request) >> { Optional.of(expectedResponse) }
-    response == expectedResponse
-  }
-
-  static Response mockResponse() {
-    new Response("url", 200, "nothing", emptyList(), new TypedByteArray("application/json", "response".bytes))
   }
 }
