+++ echo-notifications/src/test/groovy/com/netflix/spinnaker/echo/slack/SlackNotificationServiceSpec.groovy
@@ -16,27 +16,18 @@
 
 package com.netflix.spinnaker.echo.slack
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.netflix.spinnaker.echo.api.Notification
-import com.netflix.spinnaker.echo.api.Notification.InteractiveActionCallback
-import com.netflix.spinnaker.echo.config.SlackConfig
+import com.netflix.spinnaker.echo.config.SlackLegacyProperties
 import com.netflix.spinnaker.echo.notification.NotificationTemplateEngine
-import com.netflix.spinnaker.kork.web.exceptions.InvalidRequestException
-import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpMethod
-import org.springframework.http.RequestEntity
 import spock.lang.Specification
 import spock.lang.Subject
 
 class SlackNotificationServiceSpec extends Specification {
 
   def slack = Mock(SlackService)
-  def slackHookService = Mock(SlackConfig.SlackHookService)
   @Subject def service = new SlackNotificationService (
     slack,
-    slackHookService,
-    new NotificationTemplateEngine(),
-    new ObjectMapper()
+    new NotificationTemplateEngine()
   )
 
   def "supports the SLACK notification type"() {
@@ -55,7 +46,7 @@ class SlackNotificationServiceSpec extends Specification {
     notification.severity = Notification.Severity.NORMAL
     notification.additionalContext["body"] = "text"
 
-    slack.configProperties >> new SlackConfig.SlackProperties()
+    slack.config >> new SlackLegacyProperties()
 
     slack.sendMessage(*_) >> { message, channel, asUser -> }
 
@@ -64,117 +55,5 @@ class SlackNotificationServiceSpec extends Specification {
 
     then:
     notification.to.size() * slack.sendMessage(*_)
-  }
-
-  def "parsing an interaction callback from Slack returns the matching generic representation"() {
-    given:
-    String slackRequestBody = "payload=" +
-      URLEncoder.encode(getClass().getResource("/slack/callbackRequestBody.txt").text)
-
-    RequestEntity<String> request = new RequestEntity<>(
-      slackRequestBody, new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    slack.verifyToken(*_) >> { }
-    slack.getUserInfo(*_) >> new SlackService.SlackUserInfo(email: "john@doe.com")
-
-    InteractiveActionCallback expectedCallbackObject = new InteractiveActionCallback(
-      actionPerformed: new Notification.ButtonAction(
-        name: "approval",
-        value: "yes"
-      ),
-      serviceId: "keel",
-      messageId: "01DYNYK94DE4901K9P802DB4ZT",
-      user: "john@doe.com"
-    )
-
-    when:
-    InteractiveActionCallback callbackObject = service.parseInteractionCallback(request)
-
-    then:
-    callbackObject == expectedCallbackObject
-  }
-
-  def "failing to retrieve user info falls back to user name when parsing an interaction callback from Slack"() {
-    given:
-    String slackRequestBody = "payload=" +
-      URLEncoder.encode(getClass().getResource("/slack/callbackRequestBody.txt").text)
-
-    RequestEntity<String> request = new RequestEntity<>(
-      slackRequestBody, new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    slack.verifyToken(*_) >> { }
-    slack.getUserInfo(*_) >> { throw new Exception("oops!") }
-
-    InteractiveActionCallback expectedCallbackObject = new InteractiveActionCallback(
-      actionPerformed: new Notification.ButtonAction(
-        name: "approval",
-        value: "yes"
-      ),
-      serviceId: "keel",
-      messageId: "01DYNYK94DE4901K9P802DB4ZT",
-      user: "john.doe"
-    )
-
-    when:
-    InteractiveActionCallback callbackObject = service.parseInteractionCallback(request)
-
-    then:
-    callbackObject == expectedCallbackObject
-  }
-
-
-  def "parsing a malformed interaction callback from Slack throws an exception"() {
-    given:
-    String slackRequestBody = "content=suspicious"
-    RequestEntity<String> request = new RequestEntity<>(
-      slackRequestBody, new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    when:
-    service.parseInteractionCallback(request)
-
-    then:
-    thrown(InvalidRequestException)
-  }
-
-  def "failing to verify the token from Slack throws an exception"() {
-    given:
-    String slackRequestBody = "payload=" +
-      URLEncoder.encode(getClass().getResource("/slack/callbackRequestBody.txt").text)
-
-    RequestEntity<String> request = new RequestEntity<>(
-      slackRequestBody, new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    slack.verifyToken(*_) >> { throw new InvalidRequestException() }
-    slack.getUserInfo(*_) >> { }
-
-    when:
-    service.parseInteractionCallback(request)
-
-    then:
-    thrown(InvalidRequestException)
-  }
-
-  def "responding to a callback calls Slack with the response_url and replaces the buttons in the original message with the clicked action"() {
-    given:
-    String payload = getClass().getResource("/slack/callbackRequestBody.txt").text
-    String slackRequestBody = "payload=" + URLEncoder.encode(payload, "UTF-8")
-
-    RequestEntity<String> request = new RequestEntity<>(
-      slackRequestBody, new HttpHeaders(), HttpMethod.POST, new URI("/notifications/callbaks"))
-
-    slack.verifyToken(*_) >> { }
-    slack.getUserInfo(*_) >> { }
-
-    Map parsedPayload = new ObjectMapper().readValue(payload, Map)
-    Map originalMessage = parsedPayload.original_message
-    Map expectedResponse = new HashMap(originalMessage)
-    expectedResponse.attachments[0].remove("actions")
-    expectedResponse.attachments[0].text += "\n\nUser <@${parsedPayload.user.id}> clicked the *Approve* action."
-
-    when:
-    service.respondToCallback(request)
-
-    then:
-    1 * slackHookService.respondToMessage("/actions/T00000000/0123456789/abcdefgh1234567", expectedResponse)
   }
 }
